+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added cost center custom field to the settings backend
+
 ## [1.6.0] - 2023-04-17
 ### Added
 - Addded customFields to the b2bCheckoutSettings object